backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Celery
celery[redis]==5.3.6
redis==5.0.2
croniter==2.0.1

# Auth
ldap3==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Docker
docker==7.0.0

# Monitoring
psutil==5.9.8

# Email
aiosmtplib==3.0.1

# Utils
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.27.0
websockets==12.0
